requirements.txt
streamlit
pinecone-client
cohere
sentence-transformers
pdfplumber
PyPDF2
torch
transformers
numpy
pandas
nltk